Avoid warnings with recent update of Lua.

+++ lua/src/lparser.c
@@ -838,7 +838,7 @@ static BinOpr subexpr (LexState *ls, expdesc *v, unsigned int limit) {
   else simpleexp(ls, v);
   /* expand while operators have priorities higher than `limit' */
   op = getbinopr(ls->t.token);
-  while (op != OPR_NOBINOPR && priority[op].left > limit) {
+  while (op != OPR_NOBINOPR && priority[op].left > (int)limit) {
     expdesc v2;
     BinOpr nextop;
     luaX_next(ls);
